saving during armagedon
