+++ src/app/analysis/analysis.component.ts
@@ -1,26 +1,21 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, inject } from "@angular/core";
 import { Quote } from "../analysis/analysis";
+import { SelectedcvsService } from "../services/selectedcvs.service";
 
 @Component({
   standalone: true,
-  selector: 'app-score',
-  templateUrl: './analysis.component.html',
-  styleUrls: ['./analysis.component.css']
+  selector: "app-score",
+  templateUrl: "./analysis.component.html",
+  styleUrls: ["./analysis.component.css"],
+  providers: [SelectedcvsService]
 })
-
 export class AnalysisComponent implements OnInit {
-  title = 'HTTP using native fetch API';
-  url: string = 'https://type.fit/api/quotes';
+  title = "HTTP using native fetch API";
+  url: string = "https://type.fit/api/quotes";
   quotes: Quote[] = [];
+  data = inject(SelectedcvsService);
 
   ngOnInit() {
-    console.log("coucou je suis oninit")
-    fetch(this.url)
-      .then(response => response.json())
-      .then((data : Quote[]) => {
-        this.quotes=data;
-      })
-      .then(() => console.log("ceci est fetch", this.quotes));
+    console.log(this.data.likedCVs);
   }
 }
-
