+++ src/app/evaluation/evaluation.component.html
@@ -43,7 +43,7 @@
           </ul>
         </section>
         <section class="button-section">
-          <button class="button-accent button-accent3" (click)="onDislike(cv.id)" [disabled]="likedCVs.includes(cv.id)">Ecarter</button>
+          <button class="button-accent button-accent3" (click)="onDislike(cv.id)" [disabled]="data.likedCVs.includes(cv.id)">Ecarter</button>
           <button class="button-accent button-accent2" (click)="onLike(cv.id)" [disabled]="dislikedCVs.includes(cv.id)">Retenir</button>
         </section>
       </article>
