+++ src/app/evaluation/evaluation.component.ts
@@ -1,5 +1,6 @@
-import { Component } from "@angular/core";
+import { Component, inject } from "@angular/core";
 import evaluation from "./evaluation1.json";
+import { SelectedcvsService } from "../services/selectedcvs.service";
 
 @Component({
   selector: "app-evaluation",
@@ -7,12 +8,14 @@ import evaluation from "./evaluation1.json";
   imports: [],
   templateUrl: "./evaluation.component.html",
   styleUrl: "./evaluation.component.css",
+  providers: [SelectedcvsService]
 })
 export class EvaluationComponent {
   cvs = evaluation.cvs;
   jobDescription = evaluation.jobDescription;
   dislikedCVs: number[] = [];
-  likedCVs: number[] = [];
+
+  data = inject(SelectedcvsService);
 
   onDislike(id: number): void {
     if (!this.dislikedCVs.includes(id)) {
@@ -23,17 +26,10 @@ export class EvaluationComponent {
         this.dislikedCVs.splice(index, 1);
       }
     }
-    console.log("dislikedCVs : ", this.dislikedCVs);
   }
 
   onLike(id: number): void {
-    if (!this.likedCVs.includes(id)) {
-      this.likedCVs.push(id);
-    } else {
-      const index = this.likedCVs.indexOf(id);
-      if (index > -1) {
-        this.likedCVs.splice(index, 1);
-      }
-    }
+    this.data.add(id);
+    console.log(this.data.likedCVs);
   }
 }
